# api_gateway/docs/swagger.yaml
definitions:
  utils.AddAddress:
    properties:
      city:
        type: string
      country:
        type: string
      housename:
        type: string
      pincode:
        type: string
      state:
        type: string
      street:
        type: string
    required:
    - city
    - country
    - housename
    - pincode
    - state
    - street
    type: object
...
  utils.BodyLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  utils.BodySignUpuser:
    properties:
      confirmpassword:
        type: string
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      mobilenum:
        maxLength: 10
        minLength: 10
        type: string
      password:
        type: string
      referalcode:
        type: string
    required:
    - confirmpassword
    - email
    - firstname
    - lastname
    - mobilenum
    - password
    type: object
  utils.Otpverify:
    properties:
      newpassword:
        type: string
      otp:
        type: string
      otpid:
        type: string
      referalcode:
        type: string
    required:
    - otp
    type: object
  utils.Response:
    properties:
      response:
        type: string
    type: object
host: localhost:4000
info:
  contact: {}
  description: SPORTZONE_E-COMMERCE MICROSERVICE API built using Go, PSQL, REST API
    following Clean Architecture.
  title: E-COMMERCE MICROSERVICE
  version: "2.0"
paths:
...
  /admin/login:
    post:
      consumes:
      - application/json
      description: VERIFY THE EMAIL,PASSWORD AND GENERATE A JWT TOKEN AND SET IT TO
        A COOKIE
      operationId: ADMIN-LOGIN
      parameters:
      - description: Enter the email and password
        in: body
        name: login_details
        required: true
        schema:
          $ref: '#/definitions/utils.BodyLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR ADMIN LOGIN
      tags:
      - ADMIN
  /admin/logout:
    post:
      consumes:
      - application/json
      description: ADMIN LOGOUT
      operationId: ADMIN-LOGOUT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR ADMIN LOGOUT
      tags:
      - ADMIN
...
  /admin/signup:
    post:
      consumes:
      - application/json
      description: CREATE A NEW ADMIN WITH REQUIRED DETAILS
      operationId: SIGNUP-ADMIN
      parameters:
      - description: New Admin Details
        in: body
        name: admin_details
        schema:
          $ref: '#/definitions/utils.BodySignUpuser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR NEW ADMIN SIGN UP
      tags:
      - ADMIN
  /user/cart:
    get:
      consumes:
      - application/json
      description: LISTING CART AND ITEMS FROM USERS END
      operationId: USER-LIST-CART
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR DISPLAYING CART TO USER
      tags:
      - USER
  /user/cart/add:
    put:
      consumes:
      - application/json
      description: ADDING ITEMS TO CART FROM USERS END
      operationId: USER-ADD-TO-CART
      parameters:
      - description: Enter the product id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR ADDING PRODUCTS TO CART BY USER
      tags:
      - USER
...
  /user/login:
    post:
      consumes:
      - application/json
      description: VERIFY THE EMAIL,PASSWORD AND GENERATE A JWT TOKEN AND SET IT TO
        A COOKIE
      operationId: USER-LOGIN
      parameters:
      - description: Enter the email and password
        in: body
        name: login_details
        required: true
        schema:
          $ref: '#/definitions/utils.BodyLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR USER LOGIN
      tags:
      - USER
  /user/logout:
    post:
      consumes:
      - application/json
      description: LOGOUT USER AND ALSO CLEAR COOKIES
      operationId: USER-LOGOUT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR USER LOGOUT
      tags:
      - USER
...
  /user/profile/address:
    get:
      consumes:
      - application/json
      description: LISTING ALL ADDRESSES FOR THE PARTICULAR USER
      operationId: USER-LIST-ADDRESS
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR LISTING ADDRESSES
      tags:
      - USER
  /user/profile/address/add:
    post:
      consumes:
      - application/json
      description: ADDING NEW ADDRESS TO USER PROFILE
      operationId: USER-ADD-ADDRESS
      parameters:
      - description: Add the address details
        in: body
        name: address_details
        required: true
        schema:
          $ref: '#/definitions/utils.AddAddress'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR ADDING ADDRESS
      tags:
      - USER
  /user/signup:
    post:
      consumes:
      - application/json
      description: CREATE A NEW USER WITH REQUIRED DETAILS
      operationId: SIGNUP-USER
      parameters:
      - description: New user Details
        in: body
        name: user_details
        schema:
          $ref: '#/definitions/utils.BodySignUpuser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR NEW USER SIGN UP
      tags:
      - USER
  /user/signup/otp/verify:
    post:
      consumes:
      - application/json
      description: VERIFY THE OTP AND UPDATE THE VERIFIED COLUMN
      operationId: SIGNUP-USER-OTP-VERIFY
      parameters:
      - description: otp
        in: body
        name: otp_details
        schema:
          $ref: '#/definitions/utils.Otpverify'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR NEW USER SIGN UP OTP VERIFICATION
      tags:
      - USER
swagger: "2.0"